Add files via upload
It's an app for safety purpose... It has multiple safety features which provide enhanced security with live cam which can work even after the screen is off. It alse have feature to track or find nearby Police Station and Alert Siren. Live Location sharing And safe walk Feature which detect and Trace whether you are walking or not and so on.